Create MatchingGame
Next, we'll define a MatchingGame class to represent the matching game. The game will have a list of words to match, and a method to check whether the user has matched all of the words correctly.

Finally, we'll define a main function to run the matching game. The game will prompt the user to enter a translation, and will keep track of how many attempts the user has made.

This is just a basic example, but it should give you an idea of how you can create a language-learning app and a matching game minigame using basic Python functions and data structures. You can modify the MatchingGame class to handle multiple attempts per word, or to shuffle the words before each game to make it more challenging.